Adding a button to create a new card element

+++ src/components/ReusableComponents/ModalOfItem/ConfirmModal.tsx
@@ -13,7 +13,7 @@ interface IConfirmModal {
 
 const ConfirmModal = ({ openModal, handleCloseModal, handleClose, item, deleteItem }: IConfirmModal) => {
   const handleClickDelete = () => {
-    deleteItem(item.id).then(() => {
+    item.id && deleteItem(item.id).then(() => {
       handleCloseModal();
       handleClose();
     });
